+++ contracts/interfaces/IStakingContract.sol
@@ -6,11 +6,11 @@ interface IStakingContract {
     function initialize(
         address reserveToken_,
         address tradedToken_, 
-        uint256 lockupInterval_, //  interval 
-        uint256 duration_, 
-        uint256 tradedTokenClaimFraction_, 
-        uint256 reserveTokenClaimFraction_,
-        uint256 lpClaimFraction_
+        uint64 lockupInterval_, //  interval 
+        uint64 duration_, 
+        uint64 tradedTokenClaimFraction_, 
+        uint64 reserveTokenClaimFraction_,
+        uint64 lpClaimFraction_
         ) external;
     /*
     function stake(address addr, uint256 amount) external;
